+++ src/ckanapi_harvesters/builder/__init__.py
@@ -14,11 +14,11 @@ from . import builder_resource_multi_abc
 from . import builder_resource_multi_file
 from . import mapper_datastore
 from . import builder_resource_datastore
-from . import builder_resource_multi_datastore
 from . import mapper_datastore_multi
 from . import builder_resource_datastore_multi_abc
 from . import builder_resource_datastore_multi_folder
 from . import builder_resource_datastore_file
+from . import builder_resource_multi_datastore
 from . import builder_resource_datastore_multi_harvester
 from . import builder_resource_datastore_url
 from . import builder_resource_datastore_unmanaged