+++ src/ckanapi_harvesters/harvesters/file_formats/csv_format.py
@@ -13,7 +13,7 @@ from ckanapi_harvesters.auxiliary.list_records import ListRecords
 from ckanapi_harvesters.auxiliary.ckan_auxiliary import df_download_to_csv_kwargs
 from ckanapi_harvesters.harvesters.file_formats.file_format_abc import FileFormatABC
 
-default_read_chunksize:int = 50
+default_read_chunksize: int = 1000
 
 
 class CsvFileFormat(FileFormatABC):